Delete some unnecassry code from Kalman Coating. That are especially in the parts of changing the min_fl_dist, the if statement if the direction is equal the previous one is taken out.
Two more scenarios added.

In Multiple.py the an if function is added, to check if
the maximum of cpu_core is reached and if so than it
waits until the child process are finished than it moves on.
This function helps that not to many child process are called
so the operating is not going to hang afterwards.