Multiple.py has been changed so a seed start and end can be defined.
Some changes have been done in Kalman-Coating so that the particle loop
in whole is not happening anymore. This happened only by acciden.
I did some optmizing of the code so that it does not go so many times
into for loops. This by asking for instance if particle is beside
a tile it does not need to uptdate its neighborhood list
after received a data from another particle.
This and some other changed helped to overcome the particle loop problem.
But still there is another loop problem that happens with the scenario
combined_island. This must be solved next.