almost working version of the Kalman Coating. Now the p_max_dist is getting updated, by reseting the hop count after a movement and not after a cycle.
There is now a third movement criteria.
Particle stores the direcection of where it became fl_min_dist and if the place of that direction becomes empty than the particle moves to it.

What is now not working is when there is a free location beside a particle that its distance is equal to the particles distance and the particle has a p_max_dist that is higher than its own distance, that this particle should make space for the particle with the p_max_dist.